api/requirements.txt
mysql-connector-python==8.0.33
PyJWT==2.6.0
bcrypt==4.0.1
python-dotenv==1.0.0
